# [vllm] Add support for FP8 in Triton FA kernel